+++ Classes/Data.h
@@ -8,10 +8,10 @@ using namespace std;
 class Data{
 public:
     explicit Data();
-    explicit Data(int freq, char value = '\0');
+    explicit Data(unsigned long freq, char value = '\0');
     explicit Data(char value);
 
-    int get_frequency();
+    unsigned long get_frequency();
     void set_value(char value);
 
     char get_value();
@@ -21,7 +21,7 @@ public:
     friend ostream& operator<< (ostream& out, const Data *data);
 
 private:
-    int freq;
+    unsigned long freq;
     char value;
 };
